backend/requirements.txt
uwsgi==2.0.21
aioredis==1.3.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
daphne==3.0.2
defusedxml==0.7.1
Django==4.0.3
django-channels-jwt-auth-middleware==1.0.0
django-cors-headers==3.11.0
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.0
msgpack==1.0.3
oauthlib==3.2.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.2.0
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
zope.interface==5.4.0